requirements.txt
langchain-openai
streamlit
python-dotenv
langchain_core
langchain_community
fastapi
langserve
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
transformers
sentence-transformers
langchainhub
groq
langchain_groq
beutifulsoup4
